Split storage utility functions
