+++ optuna_dashboard/_rdb_migration.py
@@ -1,4 +1,5 @@
 from __future__ import annotations
+
 import threading
 
 from bottle import Bottle
